MICROS-6935 rename platformdns autowiring
