+++ pkg/execution/svccatadmission/BUILD.bazel
@@ -20,7 +20,7 @@ go_library(
         "//pkg/execution/svccatadmission/rps:go_default_library",
         "//pkg/k8s:go_default_library",
         "//pkg/microsserver:go_default_library",
-        "//pkg/orchestration/wiring/internaldns/api:go_default_library",
+        "//pkg/orchestration/wiring/platformdns/api:go_default_library",
         "//pkg/servicecentral:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/auth:go_default_library",
@@ -54,7 +54,7 @@ go_test(
         "//pkg/execution/svccatadmission/rps/testing:go_default_library",
         "//pkg/k8s:go_default_library",
         "//pkg/microsserver:go_default_library",
-        "//pkg/orchestration/wiring/internaldns/api:go_default_library",
+        "//pkg/orchestration/wiring/platformdns/api:go_default_library",
         "//pkg/servicecatalog:go_default_library",
         "//pkg/servicecentral:go_default_library",
         "//pkg/util/auth:go_default_library",
