+++ pkg/execution/svccatadmission/internaldns.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/atlassian/voyager/pkg/k8s"
 	"github.com/atlassian/voyager/pkg/microsserver"
-	"github.com/atlassian/voyager/pkg/orchestration/wiring/internaldns/api"
+	"github.com/atlassian/voyager/pkg/orchestration/wiring/platformdns/api"
 	"github.com/atlassian/voyager/pkg/util/logz"
 	sc_v1b1 "github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
 	"github.com/pkg/errors"
@@ -19,7 +19,7 @@ import (
 )
 
 func isInternalDNSServiceClass(serviceInstance *sc_v1b1.ServiceInstance) bool {
-	return serviceInstance.Spec.ClusterServiceClassExternalID == string(apiinternaldns.ClusterServiceClassExternalID)
+	return serviceInstance.Spec.ClusterServiceClassExternalID == string(apiplatformdns.ClusterServiceClassExternalID)
 }
 
 // InternalDNSAdmitFunc checks existing DNS alias ownership via micros server API, for both InternalDNS Services and Ingress Resources
@@ -55,7 +55,7 @@ func InternalDNSAdmitFunc(ctx context.Context, microsServerClient microsServerCl
 				},
 			}, nil
 		}
-		var internalDNSSpec apiinternaldns.Spec
+		var internalDNSSpec apiplatformdns.Spec
 		if err := json.Unmarshal(serviceInstance.Spec.Parameters.Raw, &internalDNSSpec); err != nil {
 			return nil, errors.Wrap(err, "error parsing InternalDNS spec")
 		}
