+++ pkg/orchestration/wiring/kubeingress/BUILD.bazel
@@ -12,7 +12,7 @@ go_library(
         "//:go_default_library",
         "//pkg/apis/orchestration/v1:go_default_library",
         "//pkg/k8s:go_default_library",
-        "//pkg/orchestration/wiring/internaldns/api:go_default_library",
+        "//pkg/orchestration/wiring/platformdns/api:go_default_library",
         "//pkg/orchestration/wiring/wiringplugin:go_default_library",
         "//pkg/orchestration/wiring/wiringutil:go_default_library",
         "//pkg/orchestration/wiring/wiringutil/knownshapes:go_default_library",
