+++ pkg/orchestration/wiring/platformdns/api/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["types.go"],
-    importpath = "github.com/atlassian/voyager/pkg/orchestration/wiring/internaldns/api",
+    importpath = "github.com/atlassian/voyager/pkg/orchestration/wiring/platformdns/api",
     visibility = ["//visibility:public"],
     deps = [
         "//:go_default_library",
