+++ pkg/orchestration/wiring/platformdns/api/types.go
@@ -1,4 +1,4 @@
-package apiinternaldns
+package apiplatformdns
 
 import (
 	"github.com/atlassian/voyager"
@@ -6,7 +6,7 @@ import (
 )
 
 const (
-	ResourceType                  voyager.ResourceType           = "InternalDNS"
+	ResourceType                  voyager.ResourceType           = "PlatformDNS"
 	AliasTypeSimple               string                         = "Simple"
 	ClusterServiceClassExternalID servicecatalog.ClassExternalID = "f77e1881-36f3-42ce-9848-7a811b421dd7"
 	ClusterServicePlanExternalID  servicecatalog.PlanExternalID  = "0a7b1d18-cf8d-461e-ad24-ee16d3da36d3"
