+++ pkg/orchestration/wiring/registry/BUILD.bazel
@@ -10,11 +10,11 @@ go_library(
         "//pkg/orchestration/wiring/asapkey:go_default_library",
         "//pkg/orchestration/wiring/aws:go_default_library",
         "//pkg/orchestration/wiring/ec2compute/v2:go_default_library",
-        "//pkg/orchestration/wiring/internaldns:go_default_library",
-        "//pkg/orchestration/wiring/internaldns/api:go_default_library",
         "//pkg/orchestration/wiring/k8scompute:go_default_library",
         "//pkg/orchestration/wiring/k8scompute/api:go_default_library",
         "//pkg/orchestration/wiring/kubeingress:go_default_library",
+        "//pkg/orchestration/wiring/platformdns:go_default_library",
+        "//pkg/orchestration/wiring/platformdns/api:go_default_library",
         "//pkg/orchestration/wiring/postgres:go_default_library",
         "//pkg/orchestration/wiring/rds:go_default_library",
         "//pkg/orchestration/wiring/sqs:go_default_library",
