--- pkg/orchestration/wiring/testdata/internaldns.invalidparameter.error
@@ -1,1 +1,0 @@
-failed to wire up resource "internaldns" of type "InternalDNS": error invoking plugin: unsupported parameters were provided: {Target:invalid-parameter.atl-paas.net ServiceName: EnvironmentType:}
