--- pkg/orchestration/wiring/testdata/internaldns.multipleingress.error
@@ -1,1 +1,0 @@
-failed to wire up resource "internaldns" of type "InternalDNS": error invoking plugin: internaldns resources must depend on only one ingress resource
