--- pkg/orchestration/wiring/testdata/internaldns.wrongdep.error
@@ -1,1 +1,0 @@
-failed to wire up resource "internaldns" of type "InternalDNS": error invoking plugin: shape "voyager.atl-paas.net/IngressEndpointShape" is required to create ServiceBinding for "kubecompute-simple" but was not found
