+++ pkg/orchestration/wiring/testdata/platformdns.invalidparameter.error
@@ -1,0 +1,1 @@
+failed to wire up resource "platformdns" of type "PlatformDNS": error invoking plugin: unsupported parameters were provided: {Target:invalid-parameter.atl-paas.net ServiceName: EnvironmentType:}
