+++ pkg/orchestration/wiring/testdata/platformdns.multiplealias.state.yaml
@@ -13,19 +13,20 @@ spec:
         - name: microservice
           image: docker.example.com/atlassian/micros-analytics:0.1.8
   - type: KubeIngress
-    name: ingress1
+    name: ingress
     dependsOn:
       - kubecompute-simple
-  - type: KubeIngress
-    name: ingress2
-    dependsOn:
-      - kubecompute-simple
-  - type: InternalDNS
-    name: internaldns
+  - type: PlatformDNS
+    name: platformdns
     spec:
       aliases:
         - type: Simple
-          name: foo.staging.atl-paas.net
+          name: foo1.staging.atl-paas.net
+        - type: Simple
+          name: foo2.staging.atl-paas.net
+        - type: Simple
+          name: foo3.staging.atl-paas.net
+        - type: Simple
+          name: foo4.staging.atl-paas.net
     dependsOn:
-      - ingress1
-      - ingress2
+      - ingress
