+++ pkg/orchestration/wiring/testdata/platformdns.multipleresources.bundle.yaml
