+++ pkg/orchestration/wiring/testdata/platformdns.nodep.error
@@ -1,0 +1,1 @@
+failed to wire up resource "platformdns" of type "PlatformDNS": error invoking plugin: platformdns resources must depend on only one ingress resource
