+++ pkg/orchestration/wiring/testdata/platformdns.nodep.state.yaml
@@ -6,8 +6,8 @@ metadata:
   uid: 411c0040-617e-11e7-9b57-427d691976d7
 spec:
   resources:
-  - type: InternalDNS
-    name: internaldns
+  - type: PlatformDNS
+    name: platformdns
     spec:
       aliases:
         - type: Simple
